+++ common/src/main/java/dev/ryanhcode/sable/mixin/world_border/WorldBorderMixin.java
@@ -1,7 +1,7 @@
 package dev.ryanhcode.sable.mixin.world_border;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import dev.ryanhcode.sable.Sable;
-import dev.ryanhcode.sable.api.SubLevelHelper;
 import dev.ryanhcode.sable.api.sublevel.SubLevelContainer;
 import dev.ryanhcode.sable.mixinterface.world_border.WorldBorderExtension;
 import net.minecraft.core.BlockPos;
@@ -22,30 +22,26 @@ public class WorldBorderMixin implements WorldBorderExtension {
     @Unique
     private Level sable$level;
 
-    @Inject(method = "isWithinBounds(DDD)Z", at = @At("HEAD"), cancellable = true)
-    public void sable$isWithinBounds(final double x, final double z, final double offset, final CallbackInfoReturnable<Boolean> cir) {
-        if (this.sable$level == null) {
-            return;
-        }
-
+    @ModifyReturnValue(method = "isWithinBounds(DDD)Z", at = @At("RETURN"))
+    public boolean sable$isWithinBounds(final boolean original, final double x, final double z, final double offset) {
+        if (original || this.sable$level == null) return original;
         final SubLevelContainer container = SubLevelContainer.getContainer(this.sable$level);
-
-        if (container != null && container.inBounds(Mth.floor(x) >> 4, Mth.floor(z) >> 4)) {
-            cir.setReturnValue(true);
-        }
+        return container != null && container.inBounds(Mth.floor(x) >> 4, Mth.floor(z) >> 4);
     }
 
-    @Inject(method = "clampToBounds(DDD)Lnet/minecraft/core/BlockPos;", at = @At("HEAD"), cancellable = true)
-    private void sable$clampToBounds(final double x, final double y, final double z, final CallbackInfoReturnable<BlockPos> cir) {
+    @ModifyReturnValue(method = "clampToBounds(DDD)Lnet/minecraft/core/BlockPos;", at = @At("RETURN"))
+    private BlockPos sable$clampToBounds(final BlockPos original, final double x, final double y, final double z) {
         if (this.sable$level == null) {
-            return;
+            return original;
         }
 
         final SubLevelContainer container = SubLevelContainer.getContainer(this.sable$level);
 
         if (container != null && container.inBounds(Mth.floor(x) >> 4, Mth.floor(z) >> 4)) {
-            cir.setReturnValue(BlockPos.containing(x, y, z));
+            return BlockPos.containing(x, y, z);
         }
+
+        return original;
     }
 
     @Inject(method = "isInsideCloseToBorder", at = @At("HEAD"), cancellable = true)