+++ neoforge/src/main/java/dev/ryanhcode/sable/neoforge/mixin/compatibility/create/contraptions/AbstractContraptionEntityMixin.java
@@ -175,6 +175,11 @@ public abstract class AbstractContraptionEntityMixin extends Entity implements K
         }
 
         this.sable$massTracker = MassTracker.build(this.sable$blockGetter(), this.sable$localBounds);
+        if (this.sable$massTracker.isInvalid()) {
+            this.sable$massTracker = null;
+            return;
+        }
+
         final Vector3d temp = this.sable$massTracker.getCenterOfMass().negate(new Vector3d()).add(0.5, 0.5, 0.5);
         for (final FloatingBlockCluster cluster : this.sable$floatingClusterContainer.clusters) {
             cluster.getBlockData().translateOrigin(temp);